Extract Observer and Reconciler logic from controller
